improve design, decompose screen
