+++ src/test/setup.ts
@@ -1,6 +1,7 @@
 import '@testing-library/jest-dom';
-import { afterAll, afterEach, beforeAll } from 'vitest';
 import { cleanup } from '@testing-library/react';
+import { afterAll, afterEach, beforeAll } from 'vitest';
+
 import { server } from './mocks/server';
 
 beforeAll(() => {
@@ -14,4 +15,4 @@ afterEach(() => {
 
 afterAll(() => {
   server.close();
-}); 
+});
